couple code to clean up unit tests
also fixed enum encode bug. I think this approach of returning a tuple in a field's encode makes a lot of sense. I can't see when one would ever encode data from the same field multiple times (jsut do it in one go) and if you need optional data length, then add the optional flag. In terms of the bitstring, it doesn't seem that out of place to request a tuple. The only caveat i guess would be it's not clear that you're requesting a tuple so I guess make a note of that in the comments