address jack's points
also made the realization that we don't actually need an 'optional' field for data types. We do though, but with regards to parsley implementation, its not needed (but will still include). In our current implementation because the DEBUG ascii messages are all joined together (based on previous implementation) so its ASCII(string, 64) and any extra bits when you encode a string becomes \x00 (the equilvalent of leading zeros) and since \x00 isn't ever a ascii character, we can always strip away all \x00. Its useful to have because in omnibus, it'll be required to know if the ascii field is mandatory (cardinal diection) or not (debug messgaes).