Revamp parsing structure
thought about conversations I had with jack and how I would implement this in Omnibus, and I came to the conclusion that wrapping our entire CAN message in a switch would indeed be very elegant and convient in Omnibus. There is no longer any bitwise manipulation to separate msg_type and board_id, that has been entirely abstracted away with BitString (so much cleaner). However, now I have to entirely refactor test_can_message because it needs to now parse the full can message and not separately, but thats a-okay I have setup the foundational work